+++ src/ui/mod.rs
@@ -1,10 +1,10 @@
 mod app;
 mod commands;
-pub(crate) mod field_layout;
+pub mod field_layout;
 mod loading;
 mod render;
 mod tab_state;
-pub(crate) mod widgets;
+pub mod widgets;
 
 pub use app::App;
 pub use tab_state::*;